requirements.txt
django-clearcache==1.2.1
Django==4.2.2
whitenoise==6.5.0
django-pwa==1.1.0
psycopg2-binary==2.9.6
virtualenv==20.21.0
